+++ Modules/Ruleset.ComputerInfo/Public/ConvertFrom-OSBuild.ps1
@@ -87,7 +87,7 @@ function ConvertFrom-OSBuild
 }
 
 <#
-https://docs.microsoft.com/en-us/windows/release-information/
+https://docs.microsoft.com/en-us/windows/release-health/release-information
 Version	Servicing option				Availability OS build	Latest revision date	End of service
 21H1	Semi-Annual Channel				2021-05-18	19043.1165	2021-08-10	2022-12-13	2022-12-13
 20H2	Semi-Annual Channel				2020-10-20	19042.572	2020-10-13	2022-05-10	2023-05-09
@@ -110,6 +110,16 @@ Version		Servicing option					Availability OS build	Latest revision date	Mainstr
 #>
 
 Set-Variable -Name OSBuildInfo -Scope Script -Option Constant -Value ([PSCustomObject[]]@(
+		# Windows 11
+		[hashtable]@{
+			Version = "21H2"
+			Build = 22000
+		}
+		# Windows 10
+		[hashtable]@{
+			Version = "21H2"
+			Build = 19044
+		}
 		[hashtable]@{
 			Version = "21H1"
 			Build = 19043